// ui/src/lib/types/PromptTemplate.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PromptId } from "./PromptId";
import type { PromptVariable } from "./PromptVariable";
import type { StructureId } from "./StructureId";
import type { StyleId } from "./StyleId";

/**
 * Saved generation request with variable placeholders.
 */
export type PromptTemplate = { id: PromptId, name: string, text: string, variables?: Array<PromptVariable>, default_style?: StyleId | null, default_structure?: StructureId | null, };