// ui/src/lib/types/Structure.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { StructureId } from "./StructureId";
import type { StructureOutput } from "./StructureOutput";

/**
 * Canvas layout contract for AI generation.
 */
export type Structure = { id: StructureId, name: string, output: StructureOutput, layout_negatives?: string, };