// ui/src/lib/types/Style.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ModelId } from "./ModelId";
import type { Quality } from "./Quality";
import type { StyleId } from "./StyleId";

/**
 * Reusable look modifier record.
 */
export type Style = { id: StyleId, name: string, modifiers?: string, look_negatives?: string, model_pref?: ModelId | null, quality?: Quality | null, };